requirements.txt
# Core dependencies
playwright>=1.44
openai>=1.30
mcp[cli]>=1.0
pydantic>=2.7
python-dotenv>=1.0
pandas>=2.2
rich>=13.0

# Dev / test dependencies
pytest>=8.0
pytest-asyncio>=0.23
pytest-mock>=3.14
ruff>=0.4
